Remove pyswitchlib documentation and add pyswitch doc generation
Pyswitchlib code is too vast and mostly not very easy to read when it
comes to documentation. Also the documentation generation used to freeze
due to file size of the auto generated code.

The plan is to generate documenation for pyswitch wrapper, which is more
what we want users to look at and use.

There are bunch of items in pyswitch that need to be fixed and should be
fixed, including making sure that the doctests works and pass.