CANDY-723 - Fetching interface speeds to be able to persist them.
This change will enable us to identify the interface speeds, since for
SLXOs the interface speeds are not known by the type. All interfaces are
type Ethernet.

This will be used to infer, which interfaces to enable as part of device
registration, since for SLXOs the interfaces are disabled by default.