Code changes for apply_acl/remove_acl on interfaces.
This code changes is not specific to any defect though there may be some
defects fixed or introduced as part of this change. The objective of this
code change is:

1. Code segregation as per the platform. The existing code uses one
template for all the platforms (SLX, SLX_ver17s, SLX_ver_16r and NOS).
This makes the code changes difficult and affect each other even thogh the
fix is done for one platform. Increases the scope of testing.

2. During coding, it was assumed that device SLX_ver17s inherits the
configuration behavior of NOS but at many places this platform shows behaivor
similar to SLX, It make sense to respect this platform.

Points:
1. The changes are primarily tested for mac access-list as per priority.
for platforms SLX and SLX_17s.
2. NOS code is not tested yet but will be tested by tomorrow any issue will
be fixed as part of this commit.
3. Testing is done only for mac access-list, the code base should
work good for IP and IPv6 also but is not tested yet. This is done as the
priority is set for mac acl.